refactor(claude): split records, model options, and content options out of the inbound translator (split S08 L1/2)
